transformers/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
click
Flask-Migrate
python-dotenv